fully tested camera driver

+++ obc/drivers/arducam/camera_reg.h
@@ -72,7 +72,7 @@ obc_error_code_t camWriteSensorReg16_8(uint32_t regID, uint8_t regDat);
  * @param regDat Buffer to store received data
  * @return Error code indicating if the read was successful
  */
-obc_error_code_t camReadSensorReg16_8(uint8_t regID, uint8_t* regDat);
+obc_error_code_t camReadSensorReg16_8(uint32_t regID, uint8_t* regDat);
 
 /**
  * @brief Write to a list of registers over I2C
